venv_requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asciitree==0.3.3
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.3.1
Babel==2.15.0
black==24.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.1
configmypy==0.2.0
contourpy==1.2.1
cycler==0.12.1
docker-pycreds==0.4.0
docutils==0.21.2
einops==0.8.0
exceptiongroup==1.2.1
fasteners==0.19
filelock==3.13.1
flaky==3.8.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.1
h5py==3.11.0
huggingface-hub==0.24.0
idna==3.7
imagesize==1.4.1
importlib_metadata==7.2.1
importlib_resources==6.4.0
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lightning-utilities==0.11.2
lit==18.1.8
littleutils==0.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
neuraloperator==0.3.0
neurolab==0.3.5
ninja==1.11.1.1
numcodecs==0.12.1
numpy==1.26.3
numpydoc==1.7.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
ogb==1.3.6
omegaconf==2.3.0
opt-einsum==3.3.0
outdated==0.2.2
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pathtools==0.1.2
pillow==10.2.0
platformdirs==4.2.2
pluggy==1.5.0
protobuf==4.25.3
psutil==6.0.0
pyarrow==16.1.0
pycodestyle==2.12.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pytest==8.2.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.0
pytorch-wavelet==1.2
pytorch-wavelets==1.3.0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
ray==2.31.0
rdkit-pypi==2022.9.5
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentry-sdk==2.6.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sympy==1.12
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorly==0.8.1
tensorly-torch==0.5.0
threadpoolctl==3.5.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.4.0
torch-harmonics==0.6.5
torchaudio==2.4.0
torchdata==0.7.1
torchgpipe==0.0.7
torchmetrics==1.4.0.post0
torchpippy==0.2.0
torchtext==0.18.0
torchvision==0.19.0
tqdm==4.66.4
triton==3.0.0
typer==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.2
Werkzeug==3.0.3
yarl==1.9.4
zarr==2.18.2
zipp==3.19.2
